backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
pyjwt==2.8.0
cryptography==42.0.0
python-dotenv==1.0.0
